feat: show next days weather info

+++ app/src/main/java/com/faridnia/weatherforcast/model/onecallresponse/Daily.kt
@@ -8,7 +8,7 @@ data class Daily(
     val clouds: Int,
     @SerializedName("dew_point")
     val dewPoint: Double,
-    val dt: Int,
+    val dt: Long,
     @SerializedName("feels_like")
     val feelsLike: FeelsLike,
     val humidity: Int,
